// web/lib/api-types.d.ts
export interface components {
    schemas: {
// ...
        ArticleResponse: {
            /**
             * Id
             * @default
             */
            id: string;
            /**
             * Title
             * @default
             */
            title: string;
            /**
             * Doi
             * @default
             */
            doi: string;
            /**
             * Pmid
             * @default
             */
            pmid: string;
            /**
             * Pmcid
             * @default
             */
            pmcid: string;
            /** Authors */
            authors?: string[];
            /** Year */
            year?: number | null;
            /**
             * Journal
             * @default
             */
            journal: string;
            /**
             * Abstract
             * @default
             */
            abstract: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}